requirements.txt
mcp[cli]
mesa==1.2.1
google-genai
